refactor: extract inference pattern typing
